Split started
